Bug 1741665 - Align nsCString's public size_type better with other C++ APIs. r=mccr8, a=RyanVM
Differential Revision: https://phabricator.services.mozilla.com/D131959

+++ dom/base/nsContentUtils.cpp
@@ -10603,8 +10603,7 @@ nsContentUtils::GetSubresourceCacheValidationInfo(nsIRequest* aRequest) {
   return info;
 }
 
-nsCString nsContentUtils::TruncatedURLForDisplay(nsIURI* aURL,
-                                                 uint32_t aMaxLen) {
+nsCString nsContentUtils::TruncatedURLForDisplay(nsIURI* aURL, size_t aMaxLen) {
   nsCString spec;
   if (aURL) {
     aURL->GetSpec(spec);
